+++ src/main/java/com/redhat/sast/api/model/Job.java
@@ -9,35 +9,35 @@ import jakarta.persistence.*;
 
 @Entity
 @Table(name = "job")
-public class Job {
+public class Job extends JobBase {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(name = "project_name")
-    private String projectName;
+    protected String projectName;
 
     @Column(name = "project_version")
-    private String projectVersion;
+    protected String projectVersion;
 
     @Column(name = "package_name")
-    private String packageName;
+    protected String packageName;
 
     @Column(name = "package_nvr")
-    private String packageNvr;
+    protected String packageNvr;
 
     @Column(name = "osh_scan_id")
-    private String oshScanId;
+    protected String oshScanId;
 
     @Column(name = "package_source_code_url")
     private String packageSourceCodeUrl;
 
     @Column(name = "jira_link")
-    private String jiraLink;
+    protected String jiraLink;
 
     @Column(name = "hostname")
-    private String hostname;
+    protected String hostname;
 
     @Column(name = "known_false_positives_url")
     private String knownFalsePositivesUrl;
@@ -51,22 +51,22 @@ public class Job {
 
     @Enumerated(EnumType.STRING)
     @Column(name = "status", nullable = false)
-    private JobStatus status;
+    protected JobStatus status;
 
     @Column(name = "created_at", nullable = false)
-    private LocalDateTime createdAt;
+    protected LocalDateTime createdAt;
 
     @Column(name = "started_at")
-    private LocalDateTime startedAt;
+    protected LocalDateTime startedAt;
 
     @Column(name = "completed_at")
-    private LocalDateTime completedAt;
+    protected LocalDateTime completedAt;
 
     @Column(name = "cancelled_at")
-    private LocalDateTime cancelledAt;
+    protected LocalDateTime cancelledAt;
 
     @Column(name = "tekton_url")
-    private String tektonUrl;
+    protected String tektonUrl;
 
     @Column(name = "last_updated_at")
     private LocalDateTime lastUpdatedAt;
@@ -101,68 +101,12 @@ public class Job {
         this.id = id;
     }
 
-    public String getProjectName() {
-        return projectName;
-    }
-
-    public void setProjectName(String projectName) {
-        this.projectName = projectName;
-    }
-
-    public String getProjectVersion() {
-        return projectVersion;
-    }
-
-    public void setProjectVersion(String projectVersion) {
-        this.projectVersion = projectVersion;
-    }
-
-    public String getPackageName() {
-        return packageName;
-    }
-
-    public void setPackageName(String packageName) {
-        this.packageName = packageName;
-    }
-
-    public String getPackageNvr() {
-        return packageNvr;
-    }
-
-    public void setPackageNvr(String packageNvr) {
-        this.packageNvr = packageNvr;
-    }
-
-    public String getOshScanId() {
-        return oshScanId;
-    }
-
-    public void setOshScanId(String oshScanId) {
-        this.oshScanId = oshScanId;
-    }
-
     public String getPackageSourceCodeUrl() {
         return packageSourceCodeUrl;
     }
 
     public void setPackageSourceCodeUrl(String packageSourceCodeUrl) {
         this.packageSourceCodeUrl = packageSourceCodeUrl;
-    }
-
-    public String getJiraLink() {
-        return jiraLink;
-    }
-
-    public void setJiraLink(String jiraLink) {
-        this.jiraLink = jiraLink;
-    }
-
-    public String getHostname() {
-        return hostname;
-    }
-
-    public void setHostname(String hostname) {
-        this.hostname = hostname;
     }
 
     public String getKnownFalsePositivesUrl() {
@@ -187,54 +131,6 @@ public class Job {
 
     public void setGSheetUrl(String gSheetUrl) {
         this.gSheetUrl = gSheetUrl;
-    }
-
-    public JobStatus getStatus() {
-        return status;
-    }
-
-    public void setStatus(JobStatus status) {
-        this.status = status;
-    }
-
-    public LocalDateTime getCreatedAt() {
-        return createdAt;
-    }
-
-    public void setCreatedAt(LocalDateTime createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    public LocalDateTime getStartedAt() {
-        return startedAt;
-    }
-
-    public void setStartedAt(LocalDateTime startedAt) {
-        this.startedAt = startedAt;
-    }
-
-    public LocalDateTime getCompletedAt() {
-        return completedAt;
-    }
-
-    public void setCompletedAt(LocalDateTime completedAt) {
-        this.completedAt = completedAt;
-    }
-
-    public LocalDateTime getCancelledAt() {
-        return cancelledAt;
-    }
-
-    public void setCancelledAt(LocalDateTime cancelledAt) {
-        this.cancelledAt = cancelledAt;
-    }
-
-    public String getTektonUrl() {
-        return tektonUrl;
-    }
-
-    public void setTektonUrl(String tektonUrl) {
-        this.tektonUrl = tektonUrl;
     }
 
     public LocalDateTime getLastUpdatedAt() {
